refactor(player-portal): decompose FeatPicker into focused sibling files
FeatPicker.tsx shrinks from 1109 → 218 lines. Domain logic extracted to:
- FeatFilters.tsx:      SourcePicker, SortToggle, UnmetToggle, FilterSummary
- FeatMatchRow.tsx:     FeatMatchRow, FeatMatchList (versatile-heritage split)
- FeatDetailPanel.tsx:  prereq-resolution panel (component only)
- useFeatDetail.ts:     detail-fetch hook with doc-cache + stale-cancel
- feat-bio.ts:          extractDetailBio poly-type extractor
- feat-prefetch.ts:     prefetchDocuments + resolvePrereqsForDoc

FeatPicker now delegates list+detail layout to CompendiumPicker; all
existing tests pass unchanged.
